Movido findMissingTx da nano para dentro do index
