+++ external_modules/nano/methods/index.ts
@@ -1,3 +1,2 @@
 export * from './nanoWebSocket'
 export * from './rpc'
-export * from './findMissingTx'
